updated logic for deriving default gcp sa credentials for computeService
