+++ experiments/gcp/gcp-vm-instance-stop/experiment/gcp-vm-instance-stop.go
@@ -7,7 +7,6 @@ import (
 	litmusLIB "github.com/litmuschaos/litmus-go/chaoslib/litmus/gcp-vm-instance-stop/lib"
 	"github.com/litmuschaos/litmus-go/pkg/clients"
 	"github.com/litmuschaos/litmus-go/pkg/cloud/gcp"
-	gcpCommon "github.com/litmuschaos/litmus-go/pkg/cloud/gcp/common"
 	"github.com/litmuschaos/litmus-go/pkg/events"
 	experimentEnv "github.com/litmuschaos/litmus-go/pkg/gcp/gcp-vm-instance-stop/environment"
 	experimentTypes "github.com/litmuschaos/litmus-go/pkg/gcp/gcp-vm-instance-stop/types"
@@ -102,7 +101,7 @@ func VMInstanceStop(clients clients.ClientSets) {
 	}
 
 	// Create a compute service to access the compute engine resources
-	computeService, err = gcpCommon.GetGCPComputeService()
+	computeService, err = gcp.GetGCPComputeService()
 	if err != nil {
 		log.Errorf("failed to obtain a gcp compute service, err: %v", err)
 		failStep := "[pre-chaos]: Failed to obtain a gcp compute service, err: " + err.Error()
